# spring/docs/openapi.yaml
openapi: "3.0.1"
info:
  title: "Spring部分的Swagger文档"
  description: "这是项目的Spring部分的Swagger文档"
  version: "1.0.0"
...
paths:
...
  /sql-tools/query:
    post:
      tags:
      - "SQL工具"
      summary: "执行只读SQL查询（内部）"
      description: "执行受限的只读参数化SQL查询，供FastAPI Agent远程调用"
      operationId: "executeQuery"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SqlQueryDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultSqlQueryResultVO"
...
  /sql-tools/tables:
    get:
      tags:
      - "SQL工具"
      summary: "获取表结构信息（内部）"
      description: "获取白名单内MySQL表的结构信息，供FastAPI Agent远程调用"
      operationId: "getTables"
      parameters:
      - name: "table"
        in: "query"
        required: false
        schema:
          type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultListSqlTableSchemaVO"
...
components:
  schemas:
...
    SqlQueryDTO:
      required:
      - "query"
      type: "object"
      properties:
        query:
          type: "string"
          description: "只读 SELECT 语句，使用 :param 占位"
        params:
          type: "object"
          additionalProperties:
            type: "object"
            description: "命名参数"
          description: "命名参数"
      description: "只读 SQL 查询请求"
    ResultSqlQueryResultVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          $ref: "#/components/schemas/SqlQueryResultVO"
      description: "统一响应"
    SqlQueryResultVO:
      type: "object"
      properties:
        columns:
          type: "array"
          description: "结果列名，顺序与每行值的顺序一致"
          items:
            type: "string"
            description: "结果列名，顺序与每行值的顺序一致"
        rows:
          type: "array"
          description: "结果行，每行按列顺序排列"
          items:
            type: "array"
            description: "结果行，每行按列顺序排列"
            items:
              type: "object"
              description: "结果行，每行按列顺序排列"
        row_count:
          type: "integer"
          description: "返回行数"
          format: "int32"
      description: "只读SQL查询结果"
...
    ResultListSqlTableSchemaVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          type: "array"
          description: "返回数据"
          items:
            $ref: "#/components/schemas/SqlTableSchemaVO"
      description: "统一响应"
    SqlColumnInfoVO:
      type: "object"
      properties:
        name:
          type: "string"
          description: "列名"
        type:
          type: "string"
          description: "列类型"
        key:
          type: "string"
          description: "键类型，如 PRI"
        comment:
          type: "string"
          description: "列注释"
      description: "数据库列信息"
    SqlTableSchemaVO:
      type: "object"
      properties:
        table:
          type: "string"
          description: "表名"
        row_count:
          type: "integer"
          description: "表行数，无法统计时为 -1，未查询时为空"
          format: "int32"
        columns:
          type: "array"
          description: "列信息，仅在指定表名查询时返回"
          items:
            $ref: "#/components/schemas/SqlColumnInfoVO"
      description: "数据库表结构"
...